Rely on global `dotnet` instead of installation script
For secure and compliant builds, we must use the .NET SDK installed on
the build machine, which means we cannot use the installation script any
more. Additionally, that script explicitly says not to use it for
developer machines either. While the bootstrapping cost is now slightly
higher since the developer has to manually install the SDK (and
additional runtime if necessary), this is what .NET instructs us to do.

The good news is that developers no longer need to manually add this
folder's `.dotnet` folder to the front of their `PATH` before opening VS
Code for OmniSharp to work correctly. Plus, it's simpler.